trackgen: fix double sided normals
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9473 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: e80bca4f9f393817be4061918f20e9dba886be68
Former-commit-id: 48b0dff667dc927c6452b2630b2ecccb95442a73

+++ src/tools/trackgen/ac3d.h
@@ -97,7 +97,7 @@ struct Ac3d
         }
         bool equals(const V3d &other) const
         {
-            static constexpr double  SMALL_NUM = static_cast<double>(std::numeric_limits<double>::epsilon());
+            static constexpr double  SMALL_NUM = static_cast<double>(std::numeric_limits<float>::epsilon());
 
             return std::abs(x() - other.x()) < SMALL_NUM &&
                    std::abs(y() - other.y()) < SMALL_NUM &&
